+++ src/main/java/com/studyumbackend/common/util/JwtUtil.java
@@ -45,14 +45,14 @@ public class JwtUtil {
     /**
      * 토큰에서 사용자 ID 추출
      */
-    public int getUserIdFromToken(String token) {
+    public Long getUserIdFromToken(String token) {
         Claims claims = Jwts.parser()
                 .verifyWith(getSigningKey())
                 .build()
                 .parseSignedClaims(token)
                 .getPayload();
 
-        return Integer.parseInt(claims.getSubject());
+        return Long.parseLong(claims.getSubject());
     }
 
     /**
